Moved XLink to manifest class
