+++ app/Modules/Common/MPD.php
@@ -12,7 +12,6 @@ use App\Services\Reporter\Context as ReporterContext;
 use App\Services\Reporter\TestCase;
 use Illuminate\Support\Facades\Log;
 //Module checks
-use App\Modules\Common\MPD\XLink;
 use App\Modules\Common\MPD\Schematron;
 use App\Modules\Common\MPD\XSDValidation;
 
@@ -26,9 +25,6 @@ class MPD extends Module
     public function validateMPD(): void
     {
         parent::validateMPD();
-        //IMPORTANT! We first run XlinkResolve as it generates the required files for the steps below
-        new XLink()->resolveAndValidate();
-
         new Schematron()->validateSchematron();
         new XSDValidation()->validateXSD();
     }
