+++ app/Services/MPDCache.php
@@ -11,6 +11,8 @@ use App\Services\Manifest\Period;
 use App\Services\Manifest\AdaptationSet;
 use App\Services\Manifest\Representation;
 use App\Services\Manifest\ProfileSpecificMPD;
+use Illuminate\Contracts\Filesystem\Filesystem;
+use App\Services\Manifest\XLink;
 
 class MPDCache
 {
@@ -19,12 +21,14 @@ class MPDCache
      **/
     private array $domCache = [];
     public string $error = '';
+    private Filesystem $sessionStorage;
 
     private ?\DOMDocument $document = null;
 
 
     public function __construct()
     {
+        $this->sessionStorage = session_disk();
     }
 
     public function getDocument(): \DOMDocument
@@ -67,6 +71,22 @@ class MPDCache
 
     public function getMPD(ManifestType $type = ManifestType::Regular): string
     {
+        $resolveFilename = ($type == ManifestType::Live ? "live_resolved.mpd" : "resolved.mpd");
+        $manifestFilename = ($type == ManifestType::Live ? "live_manifest.mpd" : "manifest.mpd");
+        if (!$this->sessionStorage->exists($resolveFilename)) {
+            if (!$this->sessionStorage->exists($manifestFilename)) {
+                $contents = $this->getPreviousMPD($type);
+                $this->sessionStorage->put($manifestFilename, $contents);
+            }
+            $manifest = $this->sessionStorage->get($manifestFilename);
+            $resolved = new XLink()->resolveAndValidate($manifest);
+            $this->sessionStorage->put($resolveFilename, $resolved);
+        }
+        return $this->sessionStorage->get($resolveFilename);
+    }
+
+    public function getPreviousMPD(ManifestType $type = ManifestType::Regular): string
+    {
         $cachedUrl = Cache::get(cache_path(['mpd', 'url']), '');
         if ($cachedUrl != session()->get('mpd')) {
             invalidate_mpd_cache();
@@ -77,21 +97,19 @@ class MPDCache
         Cache::remember(cache_path(['mpd','url']), 3600, function () {
             return session()->get('mpd');
         });
-        $res = Cache::remember(cache_path(['mpd','contents', $type->name]), 3600, function () use ($type) {
-            return Tracer::newSpan("Retrieve mpd")->measure(function () use ($type) {
-                $contents = '';
-                try {
-                    $contents = file_get_contents(session()->get('mpd'));
-                } catch (\Exception $e) {
-                }
-                if ($contents === false) {
-                    return '';
-                }
-                if ($type == ManifestType::Regular) {
-                    Cache::put(cache_path(['mpd','url_retrieval']), time(), $seconds = 3600);
-                }
-                return $contents;
-            });
+        $res = Tracer::newSpan("Retrieve mpd")->measure(function () use ($type) {
+            $contents = '';
+            try {
+                $contents = file_get_contents(session()->get('mpd'));
+            } catch (\Exception $e) {
+            }
+            if ($contents === false) {
+                return '';
+            }
+            if ($type == ManifestType::Regular) {
+                Cache::put(cache_path(['mpd','url_retrieval']), time(), $seconds = 3600);
+            }
+            return $contents;
         });
         if ($res == '') {
             $this->error = "Unable to retrieve MPD";
